// src/app/credit-request/models/dynamicFormFieldType.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * How the field will be rendered on the form
 */
export type DynamicFormFieldType = 'Textbox' | 'Dropdown' | 'Checkbox' | 'Radio' | 'Header' | 'Calendar';

export const DynamicFormFieldType = {
    Textbox: 'Textbox' as DynamicFormFieldType,
    Dropdown: 'Dropdown' as DynamicFormFieldType,
    Checkbox: 'Checkbox' as DynamicFormFieldType,
    Radio: 'Radio' as DynamicFormFieldType,
    Header: 'Header' as DynamicFormFieldType,
    Calendar: 'Calendar' as DynamicFormFieldType
};