// src/app/credit-request/models/dynamicFormFieldRestrictions.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Restrictions for the client-side validation
 */
export type DynamicFormFieldRestrictions = 'Alphanumeric' | 'Digits' | 'Email' | 'IBAN';

export const DynamicFormFieldRestrictions = {
    Alphanumeric: 'Alphanumeric' as DynamicFormFieldRestrictions,
    Digits: 'Digits' as DynamicFormFieldRestrictions,
    Email: 'Email' as DynamicFormFieldRestrictions,
    IBAN: 'IBAN' as DynamicFormFieldRestrictions
};