// src/app/transfers/models/api/transferModel.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface TransferModel { 
    nonce?: string;
    sourceAccountId?: string;
    destinationAccountId?: string;
    /**
     * Amount in the source account currency
     */
    amount?: number;
}